Move AddressValidator to core app
DH-1285
As it will be used in non-OMIS apps.
